+++ index.js
@@ -60,8 +60,6 @@ import FlexboxItem from './src/components/flexbox/flexbox-item.vue'
 import Flow from './src/components/flow/flow.vue'
 import FlowLine from './src/components/flow/flow-line.vue'
 import FlowState from './src/components/flow/flow-state.vue'
-import Form from './src/components/form/form.vue'
-import FormField from './src/components/form/field.vue'
 import FormatTimeFilter from './src/filters/format-time.js'
 import FormPreview from './src/components/form-preview/index.vue'
 import FriendlyTimeFilter from './src/filters/friendly-time.js'
@@ -165,6 +163,9 @@ import XProgress from './src/components/x-progress/index.vue'
 import XSwitch from './src/components/x-switch/index.vue'
 import XTable from './src/components/x-table/index.vue'
 import XTextarea from './src/components/x-textarea/index.vue'
+import XForm from './src/components/x-form/form.vue'
+import XFormField from './src/components/x-form/field.vue'
+
 
 export {
   Actionsheet,
@@ -225,8 +226,6 @@ export {
   Flow,
   FlowLine,
   FlowState,
-  Form,
-  FormField,
   FormatTimeFilter,
   FormPreview,
   FriendlyTimeFilter,
@@ -329,5 +328,7 @@ export {
   XProgress,
   XSwitch,
   XTable,
-  XTextarea
+  XTextarea,
+  XForm,
+  XFormField
 }
