--- src/components/form/components.js
@@ -1,18 +1,0 @@
-import XInput from '../x-input'
-import XTextarea from '../x-textarea'
-import XNumber from '../x-number'
-import Datetime from '../datetime'
-import PopupPicker from '../popup-picker'
-import PopupRadio from '../popup-radio'
-
-// import Uploader from 'vux-uploader-component'
-
-export default {
-  input: XInput,
-  textarea: XTextarea,
-  number: XNumber,
-  datetime: Datetime,
-  picker: PopupPicker,
-  radio: PopupRadio
-  // uploader: Uploader
-}
