--- src/components/form/index.js
@@ -1,7 +1,0 @@
-import Form from './form'
-import FormField from './field'
-
-export {
-  Form,
-  FormField
-}
