+++ src/components/map.json
@@ -59,7 +59,6 @@
   "Flow": "src/components/flow/flow.vue",
   "FlowLine": "src/components/flow/flow-line.vue",
   "FlowState": "src/components/flow/flow-state.vue",
-  "Form": "src/components/form/form.vue",
   "FormatTimeFilter": "src/filters/format-time.js",
   "FormPreview": "src/components/form-preview/index.vue",
   "FriendlyTimeFilter": "src/filters/friendly-time.js",
@@ -161,5 +160,6 @@
   "XProgress": "src/components/x-progress/index.vue",
   "XSwitch": "src/components/x-switch/index.vue",
   "XTable": "src/components/x-table/index.vue",
-  "XTextarea": "src/components/x-textarea/index.vue"
+  "XTextarea": "src/components/x-textarea/index.vue",
+  "XForm": "src/components/x-form/form.vue"
 }
