+++ src/components/x-form/index.js
@@ -1,0 +1,7 @@
+import XForm from './form'
+import XFormField from './field'
+
+export {
+  XForm,
+  XFormField
+}
