+++ src/datas/vux_component_list.json
@@ -227,30 +227,6 @@
     "hasHome": false
   },
   {
-    "name": "form",
-    "icon": "&#xe699;",
-    "items": [
-      "form",
-      "form-field"
-    ],
-    "category_en": "Form",
-    "category_zh-CN": "表单",
-    "category_order": 998,
-    "hasHome": false
-  },
-  {
-    "name": "form",
-    "icon": "&#xe699;",
-    "items": [
-      "form",
-      "form-field"
-    ],
-    "category_en": "Form",
-    "category_zh-CN": "表单",
-    "category_order": 998,
-    "hasHome": false
-  },
-  {
     "name": "grid",
     "icon": "&#xe661;",
     "items": [
@@ -701,6 +677,18 @@
   {
     "name": "x-textarea",
     "icon": "&#xe611;",
+    "category_en": "Form",
+    "category_zh-CN": "表单",
+    "category_order": 998,
+    "hasHome": false
+  },
+  {
+    "name": "x-form",
+    "icon": "&#xe699;",
+    "items": [
+      "x-form",
+      "x-form-field"
+    ],
     "category_en": "Form",
     "category_zh-CN": "表单",
     "category_order": 998,
