+++ src/demo_list.json
@@ -114,5 +114,5 @@
   "NumberRoller",
   "IconLoading",
   "InlineXSwitch",
-  "Form"
+  "XForm"
 ]
